requirements.txt
asgiref==3.7.2
astroid==3.0.1
dill==0.3.7
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==20.0.3
flake8==6.1.0
inflection==0.5.1
isort==5.12.0
mccabe==0.7.0
packaging==23.2
platformdirs==4.0.0
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pylint==3.0.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.2
typing_extensions==4.8.0
uritemplate==4.1.1
